Split styles
